+++ YandexMobileAdsExample/app/src/androidTest/java/com/yandex/ads/sample/tests/adfox/AdfoxCarouselCallbacksTest.kt
@@ -67,15 +67,10 @@ internal class AdfoxCarouselCallbacksTest : BaseUITest() {
                     "Ads loaded",
                     "Impression: null",
                     "Ad clicked",
-                    "Left application",
-                    "Returned to application",
                     "Impression: null",
                     "Ad clicked",
-                    "Left application",
-                    "Returned to application"
                 )
             }
         }
     }
 }
-